refactor: 풀더 구조를 계층형 구조에서 도메인 별 구조로 변경
