--- src/main/java/com/dasom/MemoReal/UserManager/domain/entity/User.java
@@ -1,38 +1,0 @@
-package com.dasom.MemoReal.UserManager.domain.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import com.fasterxml.jackson.annotation.JsonFormat;
-
-import java.time.LocalDate;
-
-@Entity
-@Table(name = "users")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long uid;
-
-    @Column(unique = true)
-    private String username;
-
-    private String email;
-    private String password;
-
-    @CreationTimestamp
-    @Column(updatable = false)
-    @JsonFormat(pattern = "yyyy-MM-dd", timezone = "Asia/Seoul") // JSON 응답 형식 지정
-    private LocalDate createdAt;
-
-    public User(String username, String email, String password) {
-        this.username = username;
-        this.email = email;
-        this.password = password;
-    }
-}
