--- src/main/java/com/dasom/MemoReal/UserManager/domain/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.dasom.MemoReal.UserManager.domain.repository;
-
-import com.dasom.MemoReal.UserManager.domain.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-    Optional<User> findByUsername(String username);
-    boolean existsByEmail(String email);
-    boolean existsByUsername(String username);
-}
