+++ src/main/java/com/dasom/MemoReal/domain/UserManager/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.dasom.MemoReal.domain.UserManager.repository;
+
+import com.dasom.MemoReal.domain.UserManager.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+    Optional<User> findByUsername(String username);
+    boolean existsByEmail(String email);
+    boolean existsByUsername(String username);
+}
